+++ lib/data/t_chain_payment_action.dart
@@ -1,22 +1,10 @@
 /// Action is what user want to do when using payment SDK
 enum TChainPaymentAction {
-  /// usecase:
+  /// Use case:
   /// - deposit your money to in-game currency
   /// - pay for an order
   deposit,
 
-  /// usecase: you want to take money out of your game wallet
+  /// Use case: user want to take money out of their game wallet
   withdraw,
-}
-
-extension TChainPaymentActionExtension on TChainPaymentAction {
-  /// Configure the path of the deep link
-  String get path {
-    switch (this) {
-      case TChainPaymentAction.deposit:
-        return 'payment_deposit';
-      case TChainPaymentAction.withdraw:
-        return 'payment_withdraw';
-    }
-  }
 }
